DROOLS-2160: refacotring MVP of new pages view
