--- drools-wb-screens/drools-wb-test-scenario-editor/drools-wb-test-scenario-editor-client/src/main/java/org/drools/workbench/screens/testscenario/client/page/audit/AuditPageView.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2017 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.drools.workbench.screens.testscenario.client.page.audit;
-
-import java.util.Set;
-
-import org.drools.workbench.models.testscenarios.shared.ExecutionTrace;
-import org.uberfire.client.mvp.UberElemental;
-
-public interface AuditPageView extends UberElemental<AuditPage> {
-
-    void showFiredRules(final ExecutionTrace executionTrace);
-
-    void showAuditLog(final Set<String> auditLogMessages);
-
-    interface Presenter {
-
-    }
-}
