--- drools-wb-screens/drools-wb-test-scenario-editor/drools-wb-test-scenario-editor-client/src/main/java/org/drools/workbench/screens/testscenario/client/page/configuration/KiePageView.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.drools.workbench.screens.testscenario.client.page.configuration;
-
-import org.drools.workbench.models.testscenarios.shared.Scenario;
-import org.drools.workbench.screens.testscenario.client.ScenarioParentWidget;
-import org.uberfire.backend.vfs.Path;
-import org.uberfire.client.mvp.UberElemental;
-
-public interface KiePageView extends UberElemental<KiePage> {
-
-    void refresh(final ScenarioParentWidget scenarioParentWidget, final Path path, final Scenario scenario);
-
-    interface Presenter {
-
-    }
-}
