+++ drools-wb-screens/drools-wb-test-scenario-editor/drools-wb-test-scenario-editor-client/src/test/java/org/drools/workbench/screens/testscenario/client/page/audit/AuditPageTest.java
@@ -36,7 +36,7 @@ import static org.mockito.Mockito.verify;
 public class AuditPageTest {
 
     @Mock
-    private AuditPageView auditPageView;
+    private AuditPage.AuditPageView auditPageView;
 
     private AuditPage auditPage;
