+++ drools-wb-screens/drools-wb-test-scenario-editor/drools-wb-test-scenario-editor-client/src/test/java/org/drools/workbench/screens/testscenario/client/page/configuration/KiePageTest.java
@@ -35,7 +35,7 @@ import static org.mockito.Mockito.verify;
 public class KiePageTest {
 
     @Mock
-    private KiePageView kiePageView;
+    private KiePage.KiePageView kiePageView;
 
     private KiePage kiePage;
